# API_DOCS.yaml
openapi: 3.0.3
info:
  title: Owami Network Testnet API
  version: "1.0"
  description: API documentation for the Owami Network blockchain MVP testnet.
servers:
  - url: http://localhost:8080/api
paths:
  /wallets/create:
    post:
      summary: Create a new wallet
      description: Generates a new wallet keypair and stores it securely.
      responses:
        '200':
          description: Wallet created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WalletResponse'
  /wallets/{address}/balance:
    get:
      summary: Get wallet balance
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Wallet balance
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BalanceResponse'
  /faucet:
    post:
      summary: Request test tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FaucetRequest'
      responses:
        '200':
          description: Tokens sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FaucetResponse'
  /transactions/send:
    post:
      summary: Send tokens
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendRequest'
      responses:
        '200':
          description: Transaction sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendResponse'
  /wallets/{address}/transactions:
    get:
      summary: Get transaction history
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransactionsResponse'
  /health:
    get:
      summary: Health check
      responses:
        '200':
          description: Service status
          content:
            text/plain:
              schema:
                type: string
                example: "OK"

components:
  schemas:
    WalletResponse:
      type: object
      properties:
        address:
          type: string
          example: "0x742d35Cc6634C0532925a3b844Bc454e4438f44e"
        privateKey:
          type: string
          example: "0x123..."
    BalanceResponse:
      type: object
      properties:
        balance:
          type: number
          example: 1000.0
    FaucetRequest:
      type: object
      required:
        - address
      properties:
        address:
          type: string
    FaucetResponse:
      type: object
      properties:
        txHash:
          type: string
        amount:
          type: number
    SendRequest:
      type: object
      required:
        - from
        - to
        - amount
        - privateKey
      properties:
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        privateKey:
          type: string
    SendResponse:
      type: object
      properties:
        txHash:
          type: string
    TransactionsResponse:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
    Transaction:
      type: object
      properties:
        hash:
          type: string
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        timestamp:
          type: string
          format: date-time